cssr_system/speech_event/speech_event_requirements.txt
nemo-toolkit[asr]==2.3.1
pyaudio==0.2.14
rospkg
rospy
torch==2.7.1
torchaudio==2.7.1
silero-vad==6.0.0
